requirements.txt
selenium==4.16.0
google-cloud-bigquery==3.15.0
webdriver-manager==4.0.1

